models::version: Inline `encodable()` method
